+++ src/main/java/net/meander/subtlyd/tags/BiomeTagsSD.java
@@ -10,6 +10,7 @@ import net.minecraft.world.level.biome.Biome;
  */
 public class BiomeTagsSD {
     public static final TagKey<Biome> IS_WINDY = create("is_windy");
+    public static final TagKey<Biome> IS_SLIGHTLY_FOGGY = create("is_slightly_foggy");
     public static final TagKey<Biome> IS_VERY_FOGGY = create("is_very_foggy");
     public static final TagKey<Biome> IS_FOGGY = create("is_foggy");
     public static final TagKey<Biome> HAS_CESPITOSE = create("has_cespitose");